Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Change-Id: I8cc78e991c5ba83ca6a58446b6e82684b9efb838
Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>